Q1 v0.2 3 dense layers, save weights
