+++ README.md
@@ -1,5 +1,5 @@
 # Big data project
 
 Questions list:
-1. Kaggle facial-keypoints-detection
+1. Kaggle: facial-keypoints-detection
 2. Large-scale video classification
